LLM-RAG/new/requirements.txt
# Requirements for BERT Tokenizer with Embeddings and Image Processing
transformers>=4.20.0
torch>=1.12.0
PyYAML>=6.0
numpy>=1.21.0

# Dependencies for image processing
Pillow>=8.0.0          # Image loading and processing
torchvision>=0.13.0    # Image transforms and preprocessing
omegaconf>=2.1.0       # Configuration management for BLIP2